# Services/Bones.Api/OpenApi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Bones API",
    "description": "Its an API and it does stuff",
    "contact": {
      "name": "GitHub Issues",
      "url": "https://github.com/CorruptComputer/Bones/issues"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/CorruptComputer/Bones/blob/develop/LICENSE"
    },
    "version": "develop"
  },
  "paths": {
...
    "/Project/projects/by-owner": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "Gets the projects for the current user, or specified organization",
        "operationId": "GetProjectsByOwnerAsync",
        "parameters": [
          {
            "name": "ownerType",
            "in": "query",
            "description": "The type of owner, User if self",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OwnershipType"
            }
          },
          {
            "name": "organizationId",
            "in": "query",
            "description": "The ID of the organization, if applicable",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Project/projects/quick-select": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "Gets the users current quick select projects",
        "operationId": "GetProjectQuickSelectAsync",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetProjectQuickSelectResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Project/{projectId}/dashboard": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "Gets a projects dashboard information",
        "operationId": "GetProjectDashboardAsync",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProjectDashboardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Project/{projectId}/settings": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "Gets a projects info by Id",
        "operationId": "GetProjectSettingsAsync",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProjectSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Project/{projectId}/fields": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "Gets the Item Fields available in a project",
        "operationId": "GetProjectItemFieldsAsync",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "The ID of the project",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProjectItemFieldsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Project/{projectId}/layouts": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "Gets the item layouts for a project, optionally filtered by the uses they are enabled for",
        "operationId": "GetProjectLayoutsAsync",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "The ID of the project",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "layoutUse",
            "in": "query",
            "description": "",
            "schema": {
              "$ref": "#/components/schemas/ItemLayoutUse"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetProjectLayoutsResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Project/{projectId}/initiatives": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "Gets the initiatives in a project",
        "operationId": "GetInitiativesInProjectAsync",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "The ID of the project",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetInitiativesInProjectResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmptyResponse": {
        "type": "object",
        "properties": {
          "workaround": {
            "type": "string",
            "description": "Without something in the body the API client will not generate anything for this type.\nDon't actually have to put anything in it.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "This is a workaround for the limitations of the API client."
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StringStringListKeyValuePair"
            },
            "description": "The errors that occurred, with the key being either the input that was invalid and the list of reasons it was invalid, or 'server' with the list of server errors.",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "description": "The error message, if any.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The response body is empty, this is a workaround for the limitations of the API client."
      },
...
      "GetProjectDashboardResponse": {
        "required": [
          "assetTypes",
          "initiativeCount",
          "initiatives",
          "projectId",
          "projectName"
        ],
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "description": "The projects ID",
            "format": "uuid"
          },
          "projectName": {
            "type": "string",
            "description": "The name of the project"
          },
          "initiativeCount": {
            "type": "integer",
            "description": "The number of initiatives in the project",
            "format": "int32"
          },
          "assetTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetTypesListModel"
            },
            "description": "The types of assets in the project"
          },
          "initiatives": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InitiativeListModel"
            },
            "description": "A list of the initiatives in the project"
          }
        },
        "additionalProperties": false,
        "description": "Response for the GetProjectDashboardAsync endpoint"
      },
...
      "GetProjectQuickSelectResponse": {
        "required": [
          "projectId",
          "projectName"
        ],
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "description": "The projects ID",
            "format": "uuid"
          },
          "projectName": {
            "type": "string",
            "description": "The name of the project"
          }
        },
        "additionalProperties": false,
        "description": "Response for the GetProjectQuickSelectAsync endpoint"
      },
...
      "ItemLayoutUse": {
        "enum": [
          "none",
          "workItems",
          "assets"
        ],
        "type": "string"
      },
...
      "OwnershipType": {
        "enum": [
          "user",
          "organization"
        ],
        "type": "string"
      },
...
    }
  }
}